worker dir + move/rename commonwealthConsumer
