+++ packages/commonwealth/server/workers/commonwealthConsumer/CommonwealthConsumer.ts
@@ -8,8 +8,8 @@ import type { BrokerConfig } from 'rascal';
 import { factory, formatFilename } from 'common-common/src/logging';
 import { RascalSubscriptions } from 'common-common/src/rabbitmq/types';
 import Rollbar from 'rollbar';
-import { RABBITMQ_URI, ROLLBAR_ENV, ROLLBAR_SERVER_TOKEN } from '../config';
-import models from '../database';
+import { RABBITMQ_URI, ROLLBAR_ENV, ROLLBAR_SERVER_TOKEN } from '../../config';
+import models from '../../database';
 import { processChainEventNotificationsCUD } from './messageProcessors/chainEventNotificationsCUDQueue';
 import { processSnapshotMessage } from './messageProcessors/snapshotConsumer';
 import { RascalConfigServices } from 'common-common/src/rabbitmq/rabbitMQConfig';
