+++ packages/commonwealth/server/workers/commonwealthConsumer/messageProcessors/chainEventNotificationsCUDQueue.ts
@@ -6,9 +6,9 @@ import {
 import { NotificationCategories } from 'common-common/src/types';
 import type { ChainEventNotification } from 'types';
 import type { Logger } from 'typescript-logging';
-import type { DB } from '../../models';
-import type { NotificationInstance } from '../../models/notification';
-import emitNotifications from '../../util/emitNotifications';
+import type { DB } from '../../../models';
+import type { NotificationInstance } from '../../../models/notification';
+import emitNotifications from '../../../util/emitNotifications';
 
 export type Ithis = {
   models: DB;
